feat: implementa EstoqueService

+++ src/main/java/com/gestaooserp/dev/repository/EstoqueRepository.java
@@ -3,5 +3,5 @@ package com.gestaooserp.dev.repository;
 import com.gestaooserp.dev.entity.Estoque;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EstoqueRepository extends JpaRepository<Estoque,Integer> {
+public interface EstoqueRepository extends JpaRepository<Estoque,Long> {
 }
